requirements.txt
flask
Pillow
redis
tensorflow
keras
opencv_python
scikit-image
